Fix various Javadoc issues
Most frequent issue was broken links due to usage of Solution_
parameter in method references. Javadoc tool cannot handle that but when
the method reference is used without parameters, the link is generated
correctly.

Another issue was with @value tag. It appears to require FQCN when
referencing a different class.

Also fixes missing period at ends of first sentences and typos.