Move setBestSolutionTimeMillis inside "if (bestScore.isSolutionInitialized())"
This change causes the solver to terminates gracefully instead of crashing if a construction heuristic
fails to initalize a solution within the Unimproved Spent Limit (if one is set).

Reason for crash before:

A construction heuristic does not update the best solution until it is finished. As such, even if
the score is improving, the Unimproved Spent Limit Termination cannot tell until the Phase is
finished (either when the solution is initalized or unimproved spent limit is passed).
As a consequence, the timer get reset at the end of the Construction Heuristic phase (and not mid-way through it).
Since the timer is reset, the Solver does not terminate and proceeds to the Local Search Phase
(if one is configured). However, the Local Search Phase required the solution to be initalized,
so a crash occurs if the Construction Heuristic could not initalize a solution within the
unimproved spent limit.

For projects that use solverConfig's, this issue does not happen often since
it is not recommended to use unimprovedSpentLimit on the entire solver.
However, for spring-boot-autoconfigure (without solverConfig.xml), you cannot set unimprovedSpentLimit
only on specific phases in application.properties, meaning you need to set
it on the solver, which can lead to this crash.

Fix:

Only update time if the solution is initalized

Alternatives:

- Redesign how Construction Heuristic update best solution
- Remove UnimprovedSpentLimit property from Spring Boot Autoconfigure