PLANNER-1996: Enable dependency check (#824)
Use maven-dependency-plugins to fail the build if any used undeclared
or unused declared dependencies are found.

The plugin is configured to ignore non-compile dependencies. This avoids
having to ignore runtime and test-runtime dependencies individually.

Explanation:
- The plugin cannot detect if a declared runtime dependency is unused.
- The plugin cannot detect if an undeclared runtime dependency is used
  transitively. It is contributor's responsibility to declare runtime
  dependencies even though they are provided transitively.
- Missing runtime dependencies should be detected by tests. It's
  contributor's responsibility to write tests using runtime dependencies
  or to document why a runtine dependency is declared in case it
  cannot be covered by tests.
- The plugin will detect unused declared compile-scope dependencies.
- The plugin will detect used undeclared (transitive) compile-scope
  dependencies.

Exception: Spring Boot Starters are designed to provided transitive
dependencies. So the OptaPlanner Spring Boot Starter is allowed to have
unused declared dependencies and projects depending on Starters are not
checked at all. The same applies for Quarkus integration modules.