PLANNER-2138 Deprecate support for JavaScript expressions in configs (#917)
This is what happens:

If the config is an integer, all JS is skipped and we return it directly. Oherwise we continue to JS code.
If a JS implementation is found, it is used. This is the usual way, and works on JDK 8 and 11.
If a JS implementation is not found, Rhino is used from the classpath. This is for JDK 15+.
Users who wish to continue using JavaScript on JDK15+ will have to add the following to their POMs:

    <dependency>
      <groupId>org.mozilla</groupId>
      <artifactId>rhino</artifactId>
    </dependency>