PLANNER-2223: Move ide-config into its own module
Before, ide-config was stored in a directory, which the
formatter plugin needed to reference in order to format the
files. This means for child modules one level down from the project
root, you need to set format.directory to "../../ide-config",
for two levels down: "../../../ide-config". This is error-prone
and not developer friendly. To fix this, we move the ide-config
into its own maven module, which can be included as a dependency
for the formatter plugin, and thus we don't need to set
format.directory. Additionally, this allows us to reuse the
ide-config in other projects without copying the files.