PLANNER-2367: Fix attempt order of class cloning in GIZMO
It appears instanceOfComparator was inverted:
a.isAssignableFrom(b) => b is more specific than a => b should be
attempted/checked in an instanceof before a. Before, in this
case, it returned -1, meaning a was before b in the list, hence
the instanceof was in reverse order! Correction was to change it
to 1. Additionally, interfaces and abstract classes are filtered
out in the Gizmo generating code (so Gizmo won't attempt to
create them).