Find undeclared @PlanningEntity subclasses in GIZMO
For instance, in TSP, an anchor might implement
a Standstill interface that is a planning entity
but declare no planning variables of its own
(which means it is not annotated with @PlanningEntity).

We still want to clone the anchor as it is a subclass
of a planning entity, but we need to know about it
at GIZMO compile time. For Quarkus, it a non-issue;
we can look into the application index (WIP -- todo).
For plain Java, it is an issue. Currently, it introspect
the PlanningSolution class' fields' type variables to
discover any hidden @PlanningEntity that declared
in a type variable.