Replace workingEntityList with entitySelector, part 2
RandomListSwapIterator is straightforward. Simply select left and right
values from (randomized) valueSelectors. Get their entity[index] coordinates
using supplies.

RandomListChangeIterator is tricky. Source part is the same as above.
Use valueSelector and supplies. I first tried to select a destination
entity using an entitySelector and than randomly select a destination
index. That worked but was unfair. So I switched back to global index
and a navigable map for the destination part. It works but solver gets
stuck on the smallest data set in a local optimum. Terminate+Solve makes
it find the global optimum. STRANGE.