Do not re-run GHA workflows when PR is marked as ready for review
That makes no sense. I often start with a draft PR. I polish it
incrementally and make sure all checks are green. Then I mark it as
ready for review. At that point I want the review to see a ready PR that
can be merged immediatelly if no issues are found. But unexpectedly,
workflows are restarted without any change in the code. That is a waste
of time (both human and machine).