PLANNER-2667: Fix consecutive collector getting corrupted on data that depends on planning variables
There was a bug where the consecutive collector gets corrupted
if its index function depends on planning variable. This is due
to how score calculation works:

1. A move is performed
2. calculateScore is called
3. The undo move is performed

Since calculateScore is not called for undo, when
the entity is later removed by the next move, its index
would have change, corrupting the consecutive collector.

To fix consecutive, the index of the value is cached on add, and
the cached index is used during remove.