Also scan EntityScan packages for Score Directors (Spring Boot)
If a user have a multi-module Spring Boot project, typically
the modules are siblings of the main app module package
(i.e. they are not a subpackage of the main app module).
The old approach only check the main app packages, which would
fail for a multi-module project where all the Solver classes
are in its own module.

There a comment saying "Does not use EntityScanner because these
classes shouldn't be found through @EntityScan", but we lack
context for why. A possible reason might be to prevent issues
if dependencies have their own score directors. If that was the
case, I think the benefit of allowing all the OptaPlanner
classes to be in their own submodule outweighs the cost of
requiring solver configuration if the user EntityScan a dependency
that have its own ScoreDirector.

The fix is fairly simple: scan the packages from both
the EntityScan annotations and the AutoConfigurationPackages.
Distinct must be used since different base packages can return
the same bean definition.