Set default phase list for problems with multiple genuine entities
Before, all problems get the phase list [ConstructionHeuristic, LocalSearch] if
the phase list was not configured in the SolverConfig. This poses an issue for
problems with multiple genuine entities:
AbstractFromConfigFactory.deduceEntityDescriptor (which is used for
ConstructionHeuristic with a null entity placer config) throws an
exception if there are multiple genuine planning entities. As a result,
if you had a problem with multiple genuine planning entities, you are
forced to specify the phase list manually.

Now, all problems get the phase list
[
    ConstructionHeuristic(defaultEntityPlacer(genuineEntity1)),
    ConstructionHeuristic(defaultEntityPlacer(genuineEntity2)),
    ...
    ConstructionHeuristic(defaultEntityPlacer(genuineEntityN)),
    LocalSearch
]
if the phase list was not configured in the SolverConfig. Since
the EntityPlacer is configured with a specific genuine entity,
no exception is thrown.  Thus, phases no longer are required to
be configured for problems with multiple genuine entities.

Additionally, deduplicated code used to generate the default
entity placer (was duplicated between QueuedEntityPlacer and
PooledEntityPlacer).

Finally, fixed a bug in the MeetingSchedulingConstraintProvider,
where it used forEachIncludingNullVars but has a join that
depended on all variables (causing a rare NPE in construction
heuristic).