PLANNER-2757 Handle null case with Drools ANC in OptaPlanner processor (#2100)
Due to a Quarkus recorder bug https://github.com/quarkusio/quarkus/issues/26889,
the droolsAlphaNetworkCompilationEnabled (DANCE) is never actually set (i.e. it is always
null) in Quarkus, meaning we need to use isDroolsAlphaNetworkCompilationEnabled(), which is computed,
instead of getDroolsAlphaNetworkCompilationEnabled(), which just returns
droolsAlphaNetworkCompilationEnabled.

This fixes the native build, which was broken because DANCE was always
null, causing it to be enabled in native mode, which DANCE is not
compatiable with.