Resolve unchecked and rawtype warnings in field cloners (#2158)
The FieldCloner class doesn't need to be generic. The type parameter is
only needed in the clone() method to ensure that the `instanceClass`,
`original` and `clone` are all of the same type.

This change makes the code simpler by removing the type parameter where
it's not needed, and resolves several unchecked and rawtype warnings.