PLANNER-2796 Add clone helper methods for interfaces and abstract classes in Gizmo (#2178)
Clone helper methods are called to get instructions on how to clone
a particular class. Special care is needed for interfaces and
abstract classes, which cannot be created directly (meaning
all instances are subclasses of the interface/abstract class).
The helper methods for interfaces and abstract base classes
call the helper method that correspond to the class of the
instance (and if it cannot find the class because it not
a planning entity, it is shallowly cloned).