Create an initial implementation of List-based K-Opt Move + Selector
The K-Opt Move and Selector is based on the code in the paper
"An Effective Implementation of K-opt Moves for the Lin-Kernighan
TSP Heuristic" by Keld Helsgaun (http://webhotel4.ruc.dk/~keld/research/LKH/KoptReport.pdf). This is a first draft that still need proper testing.

K-Opts in the paper are performed as a series of 2-Opts that are
equivalent to the K-Opt. The 2-Opts are found by finding the
signed permutation corresponding to the K-Opt, then finding the
minimal amount of signed-reversals to convert that to the identity
permutation (see 5.4 'Execution of a feasible move' in the paper
for more details). Each one of those signed-reversals corresponds
to a 2-opt move, which reverses a section of the route.

There are some Java anti-patterns in the paper:

- Arrays start at 1
- Use of goto
- Use of global variables
- Undescriptive variable names

This first draft converted all global variables into locals,
but arrays still start at 1, names were kept the same mostly,
and there is an outer loop continuation inside an inner loop.
These will need to be refactored later.

I written a basic K-Opt Selector based on sections 5.1 and 5.2
of the paper. It can be specified in the SolverConfig as
`<listKOptMoveSelector />`.