K-Opt move: rotate first element back to its original position, fix flip operations
There are several differences between the tours used in the K-Opt paper
and the tours represented by OptaPlanner's list variables. In
particular:

- In the paper, tours are undirected. In OptaPlanner, tours are
  directed.

- In the paper, tours do not have an origin. In OptaPlanner, tours
  have an orgin.

Because the tours in the paper use undirected edges, we need
to find the correct orientation of the starting and ending
edges in order to create the flips used by the K-Opts. To
accomplish this, we keep a seperate list of indices that have
the same operations performed to it, so we can look up the
current index of the endpoint when generating the move. For
instance, a 3-op on [1, 2, 3, 4, 5, 6] might consists of two
2-ops ex: 2-op(2...5) and 2-op(3...6). The first 2-op changes
the index of 3, which we need for the next 2-op.

Becauses tours in the paper do not have an origin,
[5, 1, 2, 3, 4] is considered to be the same tour
as [1, 2, 3, 4, 5]. In constraints, there is usually
a depot to first item 'imaginary' edge, meaning
[5, 1, 2, 3, 4] represents a different route than
[1, 2, 3, 4, 5] and probably have a different distance.
Thus we want the first element of the tour to be
the same, since that preserves the 'imaginary'
depot to origin edge. To do this, the k-opt
move is given the new position of the origin,
and it rotates the list so the origin is back
at the original position (for the above
example, [5, 1, 2, 3, 4] is rotated by 1
to left to become [1, 2, 3, 4, 5]).

Other Fixes:
  - Fixed off-by-1 error in calculating random k between minK and maxK

Refactors:
  - Updated names from ListKOpt... to KOptList...

  - Moved classes into list/kopt package

  - Iterate through the entire body of pickKOptMoveRec
    a couple of times to make it more likely a k-opt move
    is found.

  - Move successor, between and predeccessor to KOptListMove.

Additions:
  - Create FlipSublistMove, which flips a sublist without rotating
    it back in position. It is used in KOptListMove instead of
    TwoOptListMove (which now rotate the list back into position
    like the KOptListMove)

  - Create a new method, KOptListMove.fromRemovedAndAddedEdges(...),
    which allows creation of a new KOptListMove from the given
    lists of removed and added edges. This is the method that
    should be used to create K-Opt moves outside of the K-Opt
    move iterators, since it correctly sets all the parameters
    of the KOptDescriptor, and it easy to imagine what the
    resulting graph will look like. Do note: the edges passed
    in are undirected, meaning the resulting orientation of
    the added edges might be flipped.