Fix reversing subarray when end < start
The original code used the old sublist implementation that was
modified to work on arrays. However, this is based on an invalid
assumption: that the list and array will use the same indicies.
This is false; the forward indicies are the same, but the
reverse indicies for the first half are based on array.length,
NOT firstHalfSize. This can cause the reversal to be incorrect
if secondHalfSize < firstHalfSize. This fixes the indices and
fixes the failing test.