Lower score target in vehicle routing smoke test
When only using 2-opts, the smoke test for the
smallest dataset get stuck in a lower local optima.
Using 3-opt+ allow it to escape that local optima.
Since 3-opts and 4-opts are slower than 2-opts only,
and it only affect the smallest dataset, we lowered
the smallest datatest target score in the vehicle
routing smoke test.