PLANNER-2912: Add nearby selection support and support multiple entities for KOptListMove
Multiple entities are supported by combining the lists of each selected
entity into a single route, doing the k-opt as described in the paper +
shift, then breaking the joined route into individual routes by spliting
the route at the new location of the original endpoints. This change
allow KOptListMove to use value selectors now (since values are no
longer limited to the same entity constraint). For TwoOptListMove,
when multiple entities are used, it acts as a tail swap move, swaping
the tails of two entities (where the tail start at the given endpoint).

Because the number of values examed by the KOptListMoveIterator is
not deterministic (since it need to examine the values picked to
ensure certain properties are met), the normal nearby selection iterator
does not work (in particular, it cannot replay the origin multiple
times). To account for this, a new method "fixedNearbyOriginIterator"
was added to NearValueNearbyValueSelector, which is a nearby selection
iterator that uses a fixed origin (and thus ignores the replaying
iterator).

The KOptAffectedElements and patchCycles does not appear to work
well across multiple entities, so they are currently disabled
(causing notifications to be for a larger list section then needed
(KOptAffectedElements) and preventing generation of non-sequentiel
moves (patchCycles).