Fix the probability that a given k for a k-opt move is selected
pickedKDistribution stores the relative probability a given k
is selected for a k-opt move. A given k is generated using the
following procedure:

    pos = random.randomInt(0, sum(pickedKDistribution))
    k = minK
    while pos > pickedKDistribution[k - minK]:
        pos = pos - pickedKDistribution[k - minK]
        k++

The intent of this is to create buckets with lengths corresponding
to pickedKDistribution, and setting k based on the bucket pos falls
into. So for minK=2, maxK=4, pickedKDistribution = 3,2,1;
the intended buckets are:

k = 2: [0, 3)
k = 3: [3, 5)
k = 4: [5, 6)

However, pos > pickedKDistribution[k - minK] changes the exclusive
end bound to an inclusive one, making the buckets

k = 2: [0, 3]
k = 3: [4, 5]
k = 4: []

This effectively moves 1 pos value from the last bucket to the
first one. This prevents the final k from being picked if it
relative distribution is 1. To fix it,
pos >= pickedKDistribution[k - minK] is used as the loop guard
instead.