Re-enable patchCycles by limiting the iterator to picking values on already picked entities when certain conditions are met
Patch cycles does not work when the iterator can select values on
new entities, since by definition, those entities' routes are not
included in the combined routes of all the entities. Thus, if the
value is not in an already picked entity's route, it does not
have a cycle identifier, causing an infinite loop.

These are the conditions in which the iterator will only pick values
in already selected entities:

- During patchCycles, for the reason mentioned above.
- When a route's start or end is selected, since the
  successor and predecessor function should
  be consistent, and route endpoints successor and predecessor
  depends on the entities already selected.

By limiting the values picked by the iterator during these instances,
the KOptListMoveIterator can now patch cycles (thus allowing
non-sequential to be generated) and can include route endpoints
in moves.

Including route endpoints in move complicates the successor and
predecessor functions, which now need to take into account what
is the next and previous entities in the combined route so endpoints
can be passed.

Removed the FixedOrigin...NearbyValueIterator in favour of modifying
RandomNearbyIterator to reuse the already recorded result.